refactor(tools): split delivery meta builders
